# Fix #2864: Optimizer performance regression.