This essentially reverts commit ec3c3bd685ff32f9fbbe4e19db085cb651babf8c modulo:

- The build file changes to include the par-collections.
- An additional import to get `.toParArray` back.

Performance on testSuite/test:fastOptJS on
MacBook Air, 2.2 GHz Intel Core i7, 8 GB 1600 MHz DDR3

tl;dr: optimizer is 10x faster after the revert.

Before:

Linker: Check Infos: 5253899 us
Linker: Compute reachability: 2527607 us
Linker: Assemble LinkedClasses: 429050 us
Linker: Check IR: 2320856 us
Basic Linking: 10591499 us
Inc. optimizer: Batch mode: true
Inc. optimizer: Incremental part: 39471999 us
Inc. optimizer: Optimizing 35061 methods.
Inc. optimizer: Optimizer part: 348688137 us
Inc. optimizer: 388289830 us
Refiner: Compute reachability: 382626 us
Refiner: Assemble LinkedClasses: 82460 us
Refiner: 476345 us
Emitter: Class tree cache stats: reused: 2785 -- invalidated: 2975
Emitter: Method tree cache stats: resued: 0 -- invalidated: 24570
Emitter (write output): 6734101 us
Global IR cache stats: reused: 159 -- invalidated: 2187 -- trees read: 5747

After:

Linker: Check Infos: 5480910 us
Linker: Compute reachability: 3199914 us
Linker: Assemble LinkedClasses: 529704 us
Linker: Check IR: 2257044 us
Basic Linking: 11544069 us
Inc. optimizer: Batch mode: true
Inc. optimizer: Incremental part: 746521 us
Inc. optimizer: Optimizing 35061 methods.
Inc. optimizer: Optimizer part: 37023884 us
Inc. optimizer: 37889433 us
Refiner: Compute reachability: 380839 us
Refiner: Assemble LinkedClasses: 68282 us
Refiner: 463019 us
Emitter: Class tree cache stats: reused: 2785 -- invalidated: 2975
Emitter: Method tree cache stats: resued: 0 -- invalidated: 24570
Emitter (write output): 11078321 us
Global IR cache stats: reused: 0 -- invalidated: 2346 -- trees read: 5747