@gzm0 2017-04-06
I manually verified that this still passes:

```
> set resolvers in Global += "scala-integration" at "https://scala-ci.typesafe.com/artifactory/scala-integration/"
> ++2.13.0-pre-e2a2cba
> tools/compile
[success] Total time: 74 s, completed Apr 6, 2017 8:39:50 PM
```